Remove `env_logger`, apply `rustfmt`.
The library initialized `env_logger` (via `env_logger::init`) in library
code, which is not how it should be used. Rather the library should be
using the functions defined in the `log::` module, and the application
using the library should initialize `env_logger` so it has control over
the log output.

As `env_logger` (nor any `log` functions) are actually used at all, this
outright removes its dependency.

This commit also contains the changes of running `rustfmt` on modified
files.