Dial all peers on startup.
This allows Gossipsub to establish a proper mesh network. I don't think
this is the perfect way to go about it yet, as it seems to 'forget'
about inactive peers after a certain timeout, making it unable to
replenish mesh-network peers when they die.

Official examples all use a discovery protocol such as MDNS, with no
hint on how to go about it should you want to use a static list of
peers.

For now this should be much better than our previous setup though, were
nodes only connected to a single peer.