Increase size of instance storage.
Having it too small leads to a lot of messages ending up in the backlog,
as their instance will have already completed and have been evicted, or
will never have completed as it was evicted early.

This is undesireable as a) it causes protocol executions to fail and b)
messages stay in the backlog for up to 20 minutes, eating up a lot of
memory.